Refactor ajout de l'exception "ExceptionUtilisateurInexistant" dans le domain ainsi que la releve de cette exception lors de l'envoie d'un message dans "ServiceDiscussion"
