+++ src/main/java/app/OwLearning/Api/ErrorController.java
@@ -4,7 +4,7 @@ import app.OwLearning.Application.Exceptions.ExceptionChapitreIntrouvable;
 import app.OwLearning.Application.Exceptions.ExceptionMessageIntrouvable;
 import app.OwLearning.Application.Exceptions.ExceptionTokenInvalide;
 import app.OwLearning.Domain.Exceptions.*;
-import app.OwLearning.Shared.Exceptions.ExceptionUtilisateurInexistant;
+import app.OwLearning.Domain.Exceptions.ExceptionUtilisateurInexistant;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
